+++ railties/CHANGELOG.md
@@ -279,4 +279,8 @@
 *   Set config.action_mailer.async = true to turn on asynchronous
     message delivery *Brian Cardarella*
 
+*   Add `--open` option to `rails server` to open in default browser.
+
+    *fukayatsu*
+
 Please check [3-2-stable](https://github.com/rails/rails/blob/3-2-stable/railties/CHANGELOG.md) for previous changes.